requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
coincurve==6.0.0
colorama==0.3.9
decorator==4.0.11
Django==2.0.1
djangorestframework==3.7.7
google-api-python-client==1.6.5
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
pathlib2==2.2.1
psutil==5.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pysha3==1.0.2
pytz==2017.3
requests==2.18.4
rsa==3.4.2
scandir==1.5
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
win-unicode-console==0.5
